# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "更好的MAS游戏社区版",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/gacha": {
      "get": {
        "tags": [
          "Gacha"
        ],
        "summary": "List Records",
        "operationId": "listGachaRecords",
        "parameters": [
          {
            "name": "game",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "genshin",
                "starrail",
                "zzz",
                "wuthering",
                "arknights",
                "endfield"
              ],
              "type": "string",
              "title": "Game"
            }
          },
          {
            "name": "playerUid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Playeruid"
            }
          },
          {
            "name": "pool",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Pool"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GachaRecordsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gacha/fetch": {
      "post": {
        "tags": [
          "Gacha"
        ],
        "summary": "Fetch Records",
        "operationId": "fetchGachaRecords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GachaFetchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GachaUpdateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gacha/import": {
      "post": {
        "tags": [
          "Gacha"
        ],
        "summary": "Import Records",
        "operationId": "importGachaRecords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GachaImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GachaUpdateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gacha/export": {
      "get": {
        "tags": [
          "Gacha"
        ],
        "summary": "Export Records",
        "operationId": "exportGachaRecords",
        "parameters": [
          {
            "name": "game",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "genshin",
                "starrail",
                "zzz",
                "wuthering",
                "arknights",
                "endfield"
              ],
              "type": "string",
              "title": "Game"
            }
          },
          {
            "name": "playerUid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Playeruid"
            }
          },
          {
            "name": "fileFormat",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "bmasc",
                "uigf"
              ],
              "type": "string",
              "default": "bmasc",
              "title": "Fileformat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GachaExportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mas/start": {
      "post": {
        "tags": [
          "Mas"
        ],
        "summary": "Start",
        "operationId": "startMasTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mas/stop": {
      "post": {
        "tags": [
          "Mas"
        ],
        "summary": "Stop",
        "operationId": "stopMasTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasStopIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/kuro/sms/automatic": {
      "post": {
        "tags": [
          "KuroLogin"
        ],
        "summary": "Automatic Sms",
        "operationId": "automaticKuroSms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KuroSmsSessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KuroSmsAutomaticOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/kuro/sms/create": {
      "post": {
        "tags": [
          "KuroLogin"
        ],
        "summary": "Create Sms",
        "operationId": "createKuroSms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KuroSmsCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KuroSmsCreateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/kuro/sms/send": {
      "post": {
        "tags": [
          "KuroLogin"
        ],
        "summary": "Send Sms",
        "operationId": "sendKuroSms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KuroSmsSendIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/kuro/sms/login": {
      "post": {
        "tags": [
          "KuroLogin"
        ],
        "summary": "Login Sms",
        "operationId": "loginKuroSms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KuroSmsLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login/kuro/sms/cancel": {
      "post": {
        "tags": [
          "KuroLogin"
        ],
        "summary": "Cancel Sms",
        "operationId": "cancelKuroSms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KuroSmsSessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GachaExportOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content"
        ],
        "title": "GachaExportOut"
      },
      "GachaFetchIn": {
        "properties": {
          "game": {
            "type": "string",
            "enum": [
              "genshin",
              "starrail",
              "zzz",
              "wuthering",
              "arknights",
              "endfield"
            ],
            "title": "Game"
          },
          "source": {
            "type": "string",
            "maxLength": 16000,
            "format": "password",
            "title": "Source",
            "default": "",
            "writeOnly": true
          },
          "accountId": {
            "type": "string",
            "title": "Accountid",
            "default": ""
          },
          "playerUid": {
            "type": "string",
            "maxLength": 120,
            "title": "Playeruid",
            "default": ""
          },
          "timezone": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": -12.0,
            "title": "Timezone",
            "default": 8
          },
          "maxPages": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Maxpages",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "game"
        ],
        "title": "GachaFetchIn"
      },
      "GachaImportIn": {
        "properties": {
          "game": {
            "type": "string",
            "enum": [
              "genshin",
              "starrail",
              "zzz",
              "wuthering",
              "arknights",
              "endfield"
            ],
            "title": "Game"
          },
          "payload": {
            "type": "string",
            "maxLength": 20000000,
            "minLength": 2,
            "format": "password",
            "title": "Payload",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "game",
          "payload"
        ],
        "title": "GachaImportIn"
      },
      "GachaPoolOut": {
        "properties": {
          "poolType": {
            "type": "string",
            "title": "Pooltype"
          },
          "poolName": {
            "type": "string",
            "title": "Poolname"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "topRarity": {
            "type": "integer",
            "title": "Toprarity"
          },
          "topCount": {
            "type": "integer",
            "title": "Topcount"
          },
          "sinceTop": {
            "type": "integer",
            "title": "Sincetop"
          }
        },
        "type": "object",
        "required": [
          "poolType",
          "poolName",
          "total",
          "topRarity",
          "topCount",
          "sinceTop"
        ],
        "title": "GachaPoolOut"
      },
      "GachaRecord": {
        "properties": {
          "game": {
            "type": "string",
            "enum": [
              "genshin",
              "starrail",
              "zzz",
              "wuthering",
              "arknights",
              "endfield"
            ],
            "title": "Game"
          },
          "playerUid": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Playeruid"
          },
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Id"
          },
          "poolType": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Pooltype"
          },
          "poolName": {
            "type": "string",
            "maxLength": 180,
            "title": "Poolname",
            "default": ""
          },
          "name": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Name"
          },
          "itemId": {
            "type": "string",
            "maxLength": 100,
            "title": "Itemid",
            "default": ""
          },
          "itemType": {
            "type": "string",
            "maxLength": 80,
            "title": "Itemtype",
            "default": ""
          },
          "rarity": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Rarity"
          },
          "time": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}[ T]\\d{2}:\\d{2}:\\d{2}$",
            "title": "Time"
          },
          "timezone": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": -12.0,
            "title": "Timezone",
            "default": 8
          },
          "lang": {
            "type": "string",
            "maxLength": 20,
            "title": "Lang",
            "default": "zh-cn"
          },
          "isFree": {
            "type": "boolean",
            "title": "Isfree",
            "default": false
          },
          "kind": {
            "type": "string",
            "maxLength": 60,
            "title": "Kind",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "game",
          "playerUid",
          "id",
          "poolType",
          "name",
          "rarity",
          "time"
        ],
        "title": "GachaRecord"
      },
      "GachaRecordsOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/GachaRecord"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "players": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Players"
          },
          "pools": {
            "items": {
              "$ref": "#/components/schemas/GachaPoolOut"
            },
            "type": "array",
            "title": "Pools"
          }
        },
        "type": "object",
        "title": "GachaRecordsOut"
      },
      "GachaUpdateOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "added": {
            "type": "integer",
            "title": "Added",
            "default": 0
          },
          "fetched": {
            "type": "integer",
            "title": "Fetched",
            "default": 0
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "title": "GachaUpdateOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KuroSmsAutomaticOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "sent": {
            "type": "boolean",
            "title": "Sent",
            "default": false
          }
        },
        "type": "object",
        "title": "KuroSmsAutomaticOut"
      },
      "KuroSmsCreateIn": {
        "properties": {
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "phone": {
            "type": "string",
            "maxLength": 11,
            "minLength": 11,
            "format": "password",
            "title": "Phone",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "accountId",
          "phone"
        ],
        "title": "KuroSmsCreateIn"
      },
      "KuroSmsCreateOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "captchaId": {
            "type": "string",
            "title": "Captchaid"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin",
            "default": 600
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "captchaId"
        ],
        "title": "KuroSmsCreateOut"
      },
      "KuroSmsLoginIn": {
        "properties": {
          "sessionId": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Sessionid"
          },
          "smsCode": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "format": "password",
            "title": "Smscode",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "smsCode"
        ],
        "title": "KuroSmsLoginIn"
      },
      "KuroSmsSendIn": {
        "properties": {
          "sessionId": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Sessionid"
          },
          "verification": {
            "$ref": "#/components/schemas/GeetestV4Data"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "verification"
        ],
        "title": "KuroSmsSendIn"
      },
      "KuroSmsSessionIn": {
        "properties": {
          "sessionId": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Sessionid"
          }
        },
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "KuroSmsSessionIn"
      },
...
      "MasStartIn": {
        "properties": {
          "taskId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Taskid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "AutoProxy",
              "CycleRun"
            ],
            "title": "Mode",
            "default": "AutoProxy"
          }
        },
        "type": "object",
        "required": [
          "taskId"
        ],
        "title": "MasStartIn"
      },
      "MasStartOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "taskId": {
            "type": "string",
            "title": "Taskid"
          }
        },
        "type": "object",
        "required": [
          "taskId"
        ],
        "title": "MasStartOut"
      },
      "MasStopIn": {
        "properties": {
          "taskId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Taskid"
          }
        },
        "type": "object",
        "required": [
          "taskId"
        ],
        "title": "MasStopIn"
      },
...
      "OutBase": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          }
        },
        "type": "object",
        "title": "OutBase"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}